restructuring the project to make the code cleaner and more maintainable.
