--- final_logs.json
@@ -1,5 +1,0 @@
-[
-
-][
-{"ceramic fire balls": "Rule 5-9 Failed: 6.2:False"}
-]
