+++ inputs/config.json
@@ -4,8 +4,8 @@
   "GPT_API_KEY": "sk-proj-SXKNTOQyrqn9h7rvOKnUT3BlbkFJTeAbD5OiOZDnevt5Bq5j",
   "CONCURRENT_REQUESTS": 5,
   "RETRY_TIME": 5,
-  "PROXIES_FILE_PATH": "proxies.txt",
-  "INPUT_FILE_PATH": "terms.txt",
+  "PROXIES_FILE_PATH": "inputs/proxies.txt",
+  "INPUT_FILE_PATH": "inputs/terms.txt",
   "PAUSE_TIME": 900,
   "RULES": {
     "RULE_5_6_9": "\nRULE 5,6,9:\n\nHere are list of rules that needs to be matched against a search term. If any of the rules fails, reply with 'Rule No:False'. For example '5:False' or '6.1.4:False'\nIf all rules are satisfied, reply only with 'True'. Only write what's being asked do not add further explanations.\n\n\nRULES:\n5.\tNo products that are meant to be consumed by humans or animals (medicine, food, drinks) or skin care medicine.\n\n6.\tNo hazardous products: \n6.1. Pesticides/Pesticide devices, this includes some common examples \n6.1.1. Insecticides \n6.1.2. Rodenticides \n6.1.3. Herbicides \n6.1.4. Fungicides \n6.1.5. Repellants\n6.1.6. Antimicrobial pesticides\n6.1.7. Ultraviolet light units meant to kill pests \n6.1.8. Sound generators meant to deter pests \n6.1.9. Insect traps \n6.1.10. Ground vibrators meant to deter pests \n6.1.11. Water-treatment units, including water-treatment filters \n6.1.12. Air-treatment units, including air-treatment filters\n\n6.2. Flammable liquids \n6.3. Lithium Ion batteries\n6.4. Corrosive chemicals \n6.5. Aerosol cans \n6.6. Cleaning chemicals \n6.7. Car batteries \n6.8. Products that contain Lithium Ion batteries. If the product requires a lithium ion battery, but does not contain it when we buy it from the supplier, that's fine. \n6.9. Pressurised air/liquid canisters that are filled. Empty canisters are fine.\n\n9.    The search term can only contain english words. No Spanish, French or other languages. However model names/numbers are fine.\n\nSearch Term: {}\"\n",
