+++ inputs/rule_logs.json
@@ -1,0 +1,8 @@
+{
+    "t3 hair dryer nozzle replacement": {
+        "Rule 5-9": "Failed: "
+    },
+    "ceramic fire balls": {
+        "Rule 5-9": "Failed: "
+    }
+}
