+++ inputs/sales_rank_info/appliances.txt
