+++ inputs/sales_rank_info/beauty.txt
