+++ inputs/sales_rank_info/industrial.txt
