+++ inputs/sales_rank_info/tools.txt
