+++ inputs/terms.txt
@@ -1,0 +1,2 @@
+t3 hair dryer nozzle replacement
+ceramic fire balls
