--- rule_logs.json
@@ -1,7 +1,0 @@
-{
-    "t3 hair dryer nozzle replacement": {
-        "Rule 1": "Passed",
-        "Rule 2": "Passed",
-        "Rule 3 1": "Success: t3 hair dryer nozzle replacement."
-    }
-}
