--- terms.txt
@@ -1,1 +1,0 @@
-t3 hair dryer nozzle replacement
